backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pdfplumber==0.10.3
python-docx==1.1.0
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
Werkzeug==3.0.1
